library_api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.9
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
email-validator==2.3.0
